src/main/java/com/sia/security/user/SiaUser.java
package com.sia.security.user;

import java.util.Collection;
import java.util.List;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

public class SiaUser implements UserDetails {
	@Override
	public Collection<? extends GrantedAuthority> getAuthorities() {
		return List.of(() -> "READ");
	}

	@Override
	public String getPassword() {
		return "12345";
	}

	@Override
	public String getUsername() {
		return "sia.yoo";
	}
}
